更新 Packages_Download.sh

+++ .github/script/Packages_Download.sh
@@ -2,6 +2,7 @@
 regexper='^(https?)://[a-zA-Z0-9.-]+\.[a-zA-Z0-9]{1,}(:[0-9]{1,5})?(/[^/]*)*$'
 dir="${2}"
 function download(){
+    [[ -f "${dir}/$(basename ${1})" ]] && return
     curl -# -L --fail "${1}" -o "${dir}/$(basename ${1})"
     if [[ "$(du -b ${dir}/$(basename ${1}) 2>/dev/null | awk '{print $1}')" -le "512" ]]; then
 		echo -e "$(date '+%Y-%m-%d %H:%M:%S')\e[1;31m - 【$(basename ${1})】下载失败\e[0m"
@@ -17,7 +18,7 @@ function handle(){
 		LINE="${LINE%$'\r'}"
 		if grep -qE "$regexper" <<< "${LINE}"; then
 			echo "Downloading ${LINE}"
-			[[ $(curl -o /dev/null -s --head -w "%{http_code}" "${LINE}") -ge 400 ]] && { echo "无法访问"; exit; }
+			[[ $(curl -o /dev/null -s --head -w "%{http_code}" "${LINE}") -ge 400 ]] && { echo "无法访问"; return; }
 			download "${LINE}"
 		fi
 	done
